+++ web/src/pages/Circles/Circles.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+export default () => {
+  return (
+    <div>Circles</div>
+  )
+}
